move pollbook types to types
