+++ apps/pollbook/backend/src/index.ts
@@ -20,7 +20,6 @@ import { AvahiService } from './avahi';
 
 export type { LocalApi as Api } from './app';
 export * from './types';
-export * from './globals';
 
 loadEnvVarsFromDotenvFiles();
