+++ apps/pollbook/backend/src/receipts/mailing_address_change_receipt.tsx
@@ -1,5 +1,6 @@
 import { format } from '@votingworks/utils';
 import { assert } from '@votingworks/basics';
+import { Voter } from '@votingworks/types';
 import {
   PartyName,
   ReceiptIcon,
@@ -9,7 +10,6 @@ import {
   VoterMailingAddress,
   VoterName,
 } from './receipt_helpers';
-import { Voter } from '../types';
 
 export function MailingAddressChangeReceipt({
   voter,
