+++ apps/pollbook/backend/src/receipts/undo_check_in_receipt.tsx
@@ -1,6 +1,6 @@
 import { assert } from '@votingworks/basics';
 import { format } from '@votingworks/utils';
-import { Voter } from '../types';
+import { Voter } from '@votingworks/types';
 import {
   VoterAddress,
   StyledReceipt,
