+++ apps/pollbook/backend/src/voter_checklist.tsx
@@ -11,8 +11,8 @@ import JsBarcode from 'jsbarcode';
 import { format } from '@votingworks/utils';
 import React from 'react';
 import { iter, range } from '@votingworks/basics';
-import { Election, Precinct } from '@votingworks/types';
-import { Voter, VoterGroup } from './types';
+import { Election, Precinct, Voter } from '@votingworks/types';
+import { VoterGroup } from './types';
 
 const ROWS_PER_PAGE = 16;
