+++ libs/types/src/index.ts
@@ -43,3 +43,4 @@ export * from './ui_string_audio_clips';
 export * from './ui_string_audio_ids';
 export * from './ui_string_translations';
 export * from './ui_theme';
+export * from './pollbook';
